agent: move agent persistence code into a separate file
The persistence functions for services, service configs, checks, check
states, and agent tokens are relocated out of the huge agent/agent.go
file.
